# Conflicts:
#	CCenterRIA/UpdateDB/20250905_v127.7.0_September.sql
#	DatabaseUpdateValidator/DatabaseUpdateValidator/Nuxiba/Repository/Impl/SqlExecutor.cs
#	Jobs/CCReportsRIA_JOB.sql
#	Jobs/CCenteria_JOB.sql
#	Jobs/Master_JOB.sql
#	ccReportsRia/UpdateDB/20250905_nr146.0_Sep_Nova.sql

+++ DatabaseUpdateValidator/DatabaseUpdateValidator/Nuxiba/Helper/ISqlFileProcessor.cs
@@ -4,6 +4,6 @@
     {
         List<string> GetOrderedSqlFiles(string directoryPath, string pattern);
 
-        SortedList<double, string> MapVersionFile(List<string> sqlFiles, string pattern, double currentVersionFinal);
+        SortedList<long, string> MapVersionFile(List<string> sqlFiles, string pattern, long currentVersionFinal);
     }
 }
